+++ test/common/network/socket_option_impl_test.cc
@@ -1,30 +1,10 @@
-#include "common/network/address_impl.h"
-#include "common/network/socket_option_impl.h"
-
-#include "test/mocks/api/mocks.h"
-#include "test/mocks/network/mocks.h"
-#include "test/test_common/logging.h"
-#include "test/test_common/threadsafe_singleton_injector.h"
-
-#include "gtest/gtest.h"
-
-using testing::Invoke;
-using testing::NiceMock;
-using testing::Return;
-using testing::_;
+#include "test/common/network/socket_option_test_harness.h"
 
 namespace Envoy {
 namespace Network {
 namespace {
 
-class SocketOptionImplTest : public testing::Test {
-public:
-  SocketOptionImplTest() { socket_.local_address_.reset(); }
-
-  NiceMock<MockListenSocket> socket_;
-  Api::MockOsSysCalls os_sys_calls_;
-  TestThreadsafeSingletonInjector<Api::OsSysCallsImpl> os_calls{&os_sys_calls_};
-};
+class SocketOptionImplTest : public SocketOptionTestHarness {};
 
 // We fail to set the option if the socket FD is bad.
 TEST_F(SocketOptionImplTest, BadFd) {
@@ -37,6 +17,7 @@ TEST_F(SocketOptionImplTest, SetOptionEmptyNop) {
   SocketOptionImpl socket_option{{}, {}};
   EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
   EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PostBind));
+  EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::Listening));
 }
 
 // We fail to set the option when the underlying setsockopt syscall fails.
@@ -60,86 +41,29 @@ TEST_F(SocketOptionImplTest, SetOptionFreebindFailure) {
 // The happy path for setOption(); IP_TRANSPARENT is set to true.
 TEST_F(SocketOptionImplTest, SetOptionTransparentSuccessTrue) {
   SocketOptionImpl socket_option{true, {}};
-  if (ENVOY_SOCKET_IP_TRANSPARENT.has_value()) {
-    Address::Ipv4Instance address("1.2.3.4", 5678);
-    const int fd = address.socket(Address::SocketType::Stream);
-    EXPECT_CALL(socket_, fd()).WillRepeatedly(Return(fd));
-    EXPECT_CALL(os_sys_calls_,
-                setsockopt_(_, IPPROTO_IP, ENVOY_SOCKET_IP_TRANSPARENT.value(), _, sizeof(int)))
-        .WillOnce(Invoke([](int, int, int, const void* optval, socklen_t) -> int {
-          EXPECT_EQ(1, *static_cast<const int*>(optval));
-          return 0;
-        }));
-    EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  } else {
-    EXPECT_FALSE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  }
+  testSetSocketOptionSuccess(socket_option, IPPROTO_IP, ENVOY_SOCKET_IP_TRANSPARENT, 1,
+                             {Socket::SocketState::PreBind, Socket::SocketState::PostBind});
 }
 
 // The happy path for setOption(); IP_FREEBIND is set to true.
 TEST_F(SocketOptionImplTest, SetOptionFreebindSuccessTrue) {
   SocketOptionImpl socket_option{{}, true};
-  if (ENVOY_SOCKET_IP_FREEBIND.has_value()) {
-    Address::Ipv4Instance address("1.2.3.4", 5678);
-    const int fd = address.socket(Address::SocketType::Stream);
-    EXPECT_CALL(socket_, fd()).WillRepeatedly(Return(fd));
-    EXPECT_CALL(os_sys_calls_,
-                setsockopt_(_, IPPROTO_IP, ENVOY_SOCKET_IP_FREEBIND.value(), _, sizeof(int)))
-        .WillOnce(Invoke([](int, int, int, const void* optval, socklen_t) -> int {
-          EXPECT_EQ(1, *static_cast<const int*>(optval));
-          return 0;
-        }));
-    EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  } else {
-    EXPECT_FALSE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  }
+  testSetSocketOptionSuccess(socket_option, IPPROTO_IP, ENVOY_SOCKET_IP_FREEBIND, 1,
+                             {Socket::SocketState::PreBind});
 }
 
-// The happy path for setOpion(); IP_TRANSPARENT is set to false.
+// The happy path for setOption(); IP_TRANSPARENT is set to false.
 TEST_F(SocketOptionImplTest, SetOptionTransparentSuccessFalse) {
   SocketOptionImpl socket_option{false, {}};
-  if (ENVOY_SOCKET_IP_TRANSPARENT.has_value()) {
-    Address::Ipv4Instance address("1.2.3.4", 5678);
-    const int fd = address.socket(Address::SocketType::Stream);
-    EXPECT_CALL(socket_, fd()).WillRepeatedly(Return(fd));
-    EXPECT_CALL(os_sys_calls_,
-                setsockopt_(_, IPPROTO_IP, ENVOY_SOCKET_IP_TRANSPARENT.value(), _, sizeof(int)))
-        .Times(2)
-        .WillRepeatedly(Invoke([](int, int, int, const void* optval, socklen_t) -> int {
-          EXPECT_EQ(0, *static_cast<const int*>(optval));
-          return 0;
-        }));
-    EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-    EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PostBind));
-  } else {
-    EXPECT_FALSE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-    EXPECT_FALSE(socket_option.setOption(socket_, Socket::SocketState::PostBind));
-  }
+  testSetSocketOptionSuccess(socket_option, IPPROTO_IP, ENVOY_SOCKET_IP_TRANSPARENT, 0,
+                             {Socket::SocketState::PreBind, Socket::SocketState::PostBind});
 }
 
-// The happy path for setOpion(); IP_FREEBIND is set to false.
+// The happy path for setOption(); IP_FREEBIND is set to false.
 TEST_F(SocketOptionImplTest, SetOptionFreebindSuccessFalse) {
   SocketOptionImpl socket_option{{}, false};
-  if (ENVOY_SOCKET_IP_FREEBIND.has_value()) {
-    Address::Ipv4Instance address("1.2.3.4", 5678);
-    const int fd = address.socket(Address::SocketType::Stream);
-    EXPECT_CALL(socket_, fd()).WillRepeatedly(Return(fd));
-    EXPECT_CALL(os_sys_calls_,
-                setsockopt_(_, IPPROTO_IP, ENVOY_SOCKET_IP_FREEBIND.value(), _, sizeof(int)))
-        .WillOnce(Invoke([](int, int, int, const void* optval, socklen_t) -> int {
-          EXPECT_EQ(0, *static_cast<const int*>(optval));
-          return 0;
-        }));
-    EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  } else {
-    EXPECT_FALSE(socket_option.setOption(socket_, Socket::SocketState::PreBind));
-  }
-}
-
-// Freebind settings have no effect on post-bind behavior.
-TEST_F(SocketOptionImplTest, SetOptionFreebindPostBind) {
-  SocketOptionImpl socket_option{{}, true};
-  EXPECT_TRUE(socket_option.setOption(socket_, Socket::SocketState::PostBind));
+  testSetSocketOptionSuccess(socket_option, IPPROTO_IP, ENVOY_SOCKET_IP_FREEBIND, 0,
+                             {Socket::SocketState::PreBind});
 }
 
 // If a platform doesn't suppport IPv4 socket option variant for an IPv4 address, we fail